python/periodic_workers/app/requirements.txt
requests
schedule
numpy
pymongo
python-dateutil